requirements.txt
# Core packages
python-dotenv>=1.0.0
httpx>=0.24.1
pydantic>=2.6.1,<3.0.0
openai>=1.12.0,<2.0.0

# Notion integration
notion-client>=2.0.0

# CrewAI and related packages
crewai>=0.28.0
crewai-tools>=0.1.7

# Database
chromadb>=0.4.22

# Scheduling
schedule>=1.2.0

# Logging
python-json-logger>=2.0.7